+++ tests/conftest.py
@@ -3,7 +3,10 @@
 
 import subprocess
 
+import github3
 import pytest
+
+from oca_github_bot import config
 
 
 @pytest.fixture
@@ -29,3 +32,21 @@ def git_clone(tmp_path):
     subprocess.check_call(["git", "commit", "-m", "add somefile"], cwd=clone)
     subprocess.check_call(["git", "push", "origin", "master"], cwd=clone)
     yield clone
+
+
+@pytest.fixture(scope="module")
+def vcr_config():
+    return {
+        # Replace the Authorization request header with "DUMMY" in cassettes
+        "filter_headers": [("authorization", "DUMMY")]
+    }
+
+
+@pytest.fixture
+def gh():
+    """
+    github3 test fixture, using the configured GitHub token (if set, to
+    record vcr cassettes) or a DUMMY token (if not set, when playing back
+    cassettes).
+    """
+    return github3.login(token=(config.GITHUB_TOKEN or "DUMMY"))